refactor: unify Scene/Role/Turn types into _types.py (ENG-47)
Create src/benchflow/_types.py as the single canonical source for the
declarative Role, Scene, and Turn dataclasses.

Changes:
- New _types.py with Role (adds timeout_sec, idle_timeout_sec,
  skills_dir), Scene (adds parallel_group), and Turn
- trial.py imports from _types.py instead of defining its own copies
- _scene.py renames its internal Role to SceneRole (different fields:
  instruction, tools) with backward-compat alias
- __init__.py re-exports canonical types from _types.py; adds
  TrialRole/TrialScene backward-compat aliases
- runtime.py and trial_yaml.py updated to import from _types.py

New fields default to None — no runtime behavior changes.
